Create Boost, Simple, Type traits. Normalize configuration with interfaces
