+++ src/Predicate/Predicates/Simple/SimpleInterface.php
@@ -1,0 +1,22 @@
+<?php
+declare(strict_types = 1);
+/**
+ * MyAppIn (http://www.myappin.cz)
+ * @author    Martin Lonsky (martin@lonsky.net, +420 736 645876)
+ * @link      http://www.myappin.cz
+ * @copyright Copyright (c) MyAppIn s.r.o. (http://www.myappin.cz)
+ * Date: 25. 5. 2016
+ * Time: 22:42
+ */
+
+namespace ElasticSearchPredicate\Predicate\Predicates\Simple;
+
+
+/**
+ * Interface SimpleInterface
+ * @package ElasticSearchPredicate\Predicate\Predicates\Simple
+ */
+interface SimpleInterface {
+
+
+}
