+++ src/Predicate/Predicates/Simple/SimpleTrait.php
@@ -1,0 +1,29 @@
+<?php
+declare(strict_types = 1);
+/**
+ * MyAppIn (http://www.myappin.cz)
+ * @author    Martin Lonsky (martin@lonsky.net, +420 736 645876)
+ * @link      http://www.myappin.cz
+ * @copyright Copyright (c) MyAppIn s.r.o. (http://www.myappin.cz)
+ * Date: 25. 5. 2016
+ * Time: 22:36
+ */
+
+namespace ElasticSearchPredicate\Predicate\Predicates\Simple;
+
+
+/**
+ * Class SimpleTrait
+ * @package   ElasticSearchPredicate\Predicate\Predicates\Simple
+ * @author    Martin Lonsky (martin@lonsky.net, +420 736 645876)
+ */
+trait SimpleTrait {
+
+
+	/**
+	 * @var bool
+	 */
+	protected $_simple = true;
+
+
+}
